requirements.txt
fake-useragent
requests
websockets
colorama
pytz
tzlocal
keyboard
websocket-client
requests[socks]
pysocks